+++ octopoes/octopoes/core/service.py
@@ -5,9 +5,8 @@ from datetime import datetime, timezone
 from logging import getLogger
 from typing import overload
 
-from bits.definitions import BitDefinition, get_bit_definitions
-from bits.runner import BitRunner
-from pydantic import JsonValue, TypeAdapter
+from bits.cache import BIT_CACHE
+from bits.definitions import get_bit_definitions
 
 from octopoes.config.settings import (
     DEFAULT_LIMIT,
@@ -46,42 +45,6 @@ from octopoes.repositories.scan_profile_repository import ScanProfileRepository
 
 logger = getLogger(__name__)
 settings = Settings()
-
-
-class BitCache:
-    def __init__(self):
-        self.bit_cache: dict[BitDefinition, dict[str, tuple[list[OOI], datetime]]] = {}
-
-    def _purge(self, bit: BitDefinition):
-        now = datetime.now()
-        map(
-            self.bit_cache[bit].pop,
-            [key for key, data in self.bit_cache[bit].items() if now - data[1] <= bit.cache_lifetime],
-        )
-
-    def _bit_cache_key(self, source: OOI, parameters: list[OOI], config: dict[str, JsonValue]) -> str:
-        try:
-            serialized_ooi = str(TypeAdapter(list[OOI]).dump_json(parameters))
-            return source.model_dump_json() + serialized_ooi + json.dumps(config)
-        except Exception:
-            return str(datetime.now())
-
-    def get_bit(
-        self, bit: BitDefinition, source: OOI, parameters: list[OOI], config: dict[str, JsonValue]
-    ) -> list[OOI]:
-        key = self._bit_cache_key(source, parameters, config)
-        if bit not in self.bit_cache:
-            data = BitRunner(bit).run(source, parameters, config=config)
-            self.bit_cache[bit] = {key: (data, datetime.now())}
-        else:
-            self._purge(bit)
-            if key not in self.bit_cache[bit]:
-                data = BitRunner(bit).run(source, parameters, config=config)
-                self.bit_cache[bit][key] = (data, datetime.now())
-        return self.bit_cache[bit][key][0]
-
-
-BIT_CACHE: BitCache = BitCache()
 
 
 def find_relation_in_tree(relation: str, tree: ReferenceTree) -> list[OOI]:
@@ -619,7 +582,7 @@ class OctopoesService:
         bit_definitions = get_bit_definitions()
         for bit_id, bit_definition in bit_definitions.items():
             # loop over all oois that are consumed by the bit
-            for ooi in self.ooi_repository.faster_list_oois(
+            for ooi in self.ooi_repository.list_oois_by_object_types(
                 {bit_definition.consumes},
                 valid_time=valid_time,
             ):
